requirements.txt
Flask>=2.0
google-api-python-client>=2.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
scikit-learn>=1.0
pandas>=1.0
